refactor: dialog를 사용하여 모달 구현
