+++ src/components/BookSearchModal/BookSearchModal.module.css
@@ -2,22 +2,16 @@
   display: flex;
   flex-direction: column;
   gap: 1rem;
-  flex: 1;
-  overflow: scroll;
-}
-
-.modalWrapperSize {
   max-width: 100%;
-  width: 42rem;
-}
+  width: 42rem;}
 
 .form {
   display: flex;
   gap: 1rem;
-}
 
-.button {
-  flex: 0 0 10rem;
+  & > button {
+    flex: 0 0 10rem;
+  }
 }
 
 .bookSearchSize {
@@ -30,11 +24,12 @@
 
 .bookSearchResult {
   overflow: scroll;
+  border-radius: 0.4rem;
 
   scrollbar-width: none;
-  display: flex;
   width: 100%;
-
+  
+  display: flex;
   flex-direction: column;
   gap: 1rem;
 }
@@ -51,7 +46,7 @@
   position: relative;
   border-radius: 0.5rem;
   overflow: hidden;
-  flex: 0 0 9.6rem;
+  /* flex: 0 0 9.6rem; */
   width: 9.6rem;
   height: 14.4rem;
 
@@ -59,6 +54,7 @@
     object-fit: cover;
   }
 }
+
 
 .detailInfo {
   flex: 1;
