+++ src/components/dialog/Dialog.module.css
@@ -1,0 +1,26 @@
+.root {
+  position: fixed;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%);
+  border: 0;
+  border-radius: 0.8rem;
+
+
+  @media (max-width: 420px) {
+    height: 100%;
+    border-radius: 0;
+  }
+}
+
+.root::backdrop {
+  background: rgba(33, 33, 33, 0.4);
+}
+
+.closeButton {
+  display: block;
+  margin-left: auto;
+  font-size: 2rem;
+  width: 2rem;
+  margin-bottom: 2rem;
+}
