--- src/components/modal/ModalWrapper.module.css
@@ -1,33 +1,0 @@
-.root {
-  position: fixed;
-  top: 0;
-  left: 0;
-  bottom: 0;
-  right: 0;
-  background-color: rgba(33, 33, 33, 0.4);
-
-  display: flex;
-
-  justify-content: center;
-  align-items: center;
-}
-
-.modal {
-  display: flex;
-  flex-direction: column;
-  border-radius: 0.8rem;
-  background: var(--primary);
-  padding: 1.6rem;
-
-  @media (max-width: 420px) {
-    height: 100%;
-    border-radius: 0;
-  }
-}
-
-.closeButton {
-  font-size: 2rem;
-  width: 2rem;
-  align-self: flex-end;
-  margin-bottom: 2rem;
-}
